feat(landing): integrate login flow and rename search template
Integrate the login modal into the LandingPage and refactor the
template naming to better reflect the platform's layout hierarchy.

- Rename AppTemplate to ISearchTemplate to clarify its role as a
  workspace-specific layout.
- Integrate LoginModal into LandingPage and connect it with
  loginStore for state management.
- Update handleLaunch in LandingPage to trigger the login modal
  instead of immediate transition.
- Adjust VIcon sizing in LoginModal header to maintain visual
  consistency with the design system.
- Update ISearchPage to utilize the newly named ISearchTemplate.
